refactor: 불필요한 렌더링 최소화
